+++ CHANGELOG.md
@@ -18,6 +18,7 @@ We use the following categories for changes:
 ### Added
 - `-enable-feature=promql-per-step-stats` feature for statistics in PromQL evaluation
 - Add `readinessProbe` in helm chart [#1266]
+- Telemetry for recording rules and alerting [#1424]
 
 ### Fixed
 - Trace query returns empty result when queried with